Change "watchers_count" to "subcribers_count"
This may be a bug in the Github API, since everywhere I see "watchers_count" it's identical to "stargazers_count", which leads to duplicated numbers everywhere. However, "subscribers_count" is correct, at least on the main Repo endpoint. I'm not sure why "subscribers_count" isn't in the API for the children repos, so those are still broken. I can update the PR if I find a way to get those in. If not, we can wait for Github to fix that :)